start documentation and merge types.go/heroku.go
